requirements.txt
fastapi
uvicorn
joblib
scikit-learn
pillow
numpy
